Added linesrepresentation, dots representation now draws a circle, not a square. Still some adjustments needs to be done, as the line thickness and dot size.

+++ VisMol3/visual/shapes.py
@@ -109,7 +109,7 @@ def get_cylinder(pos, color, angle, vec_dir, length, stacks, radius=0.5, level='
     inds = copy.copy(cyd.cylinder_triangles[level])
     vertices = np.array([],dtype=np.float32)
     normals = np.array([],dtype=np.float32)
-    indices = np.array([],dtype=np.uint16)
+    indices = np.array([],dtype=np.uint32)
     colors = np.array(color*int(len(verts)/3)*stacks,dtype=np.float32)
     to_add_pos = length/float(stacks-1)
     rot_mat = mop.my_glRotatef(np.identity(4, dtype=np.float32), angle, vec_dir)[:3,:3]
